refactor: #16 파일명 및 경로 개선, 각 컨트롤러에 요청 로깅 및 응답 템플릿 적용
