+++ src/main/java/com/wholeseeds/mindle/domain/complaint/controller/ComplaintController.java
@@ -1,8 +1,9 @@
 package com.wholeseeds.mindle.domain.complaint.controller;
 
-import java.io.IOException;
 import java.util.List;
+import java.util.Map;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -11,14 +12,15 @@ import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.wholeseeds.mindle.common.code.CommonCode;
-import com.wholeseeds.mindle.common.response.ApiResponse;
-import com.wholeseeds.mindle.domain.complaint.Service.ComplaintService;
+import com.wholeseeds.mindle.common.util.ObjectUtils;
+import com.wholeseeds.mindle.domain.complaint.service.ComplaintService;
 import com.wholeseeds.mindle.domain.complaint.dto.SaveComplaintRequestDto;
 import com.wholeseeds.mindle.domain.complaint.dto.SaveComplaintResponseDto;
 import com.wholeseeds.mindle.domain.complaint.entity.Complaint;
 import com.wholeseeds.mindle.domain.complaint.exception.ImageUploadLimitExceeded;
 import com.wholeseeds.mindle.domain.complaint.mapper.ComplaintMapper;
+import com.wholeseeds.mindle.common.util.RequestLogger;
+import com.wholeseeds.mindle.common.util.ResponseTemplate;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -30,21 +32,23 @@ import lombok.extern.slf4j.Slf4j;
 public class ComplaintController {
 	private final ComplaintService complaintService;
 	private final ComplaintMapper complaintMapper;
+	private final ResponseTemplate responseTemplate;
 
 	@PostMapping(
 		value = "/save",
 		consumes = MediaType.MULTIPART_FORM_DATA_VALUE
 	)
-	public ResponseEntity<ApiResponse<SaveComplaintResponseDto>> saveComplaint(
+	public ResponseEntity<Map<String, Object>> saveComplaint(
 		@RequestPart("meta") SaveComplaintRequestDto requestDto,
-		@RequestPart(value = "image", required = false) List<MultipartFile> imageList) throws IOException {
+		@RequestPart(value = "image", required = false) List<MultipartFile> imageList
+	) {
+		RequestLogger.body(requestDto);
 
 		if (imageList.size() > 3) {
 			throw new ImageUploadLimitExceeded();
 		}
 
-		log.info("Request : {}", requestDto);
-		if (!CommonCode.objectIsNullOrEmpty(imageList)) {
+		if (!ObjectUtils.objectIsNullOrEmpty(imageList)) {
 			for (MultipartFile image : imageList) {
 				log.info("파일명 : {}", image.getOriginalFilename());
 			}
@@ -53,6 +57,6 @@ public class ComplaintController {
 		Complaint saved = complaintService.saveComplaint(requestDto, imageList);
 
 		SaveComplaintResponseDto resDto = complaintMapper.toSaveComplaintResponseDto(saved);
-		return ResponseEntity.ok(ApiResponse.ok(resDto));
+		return (responseTemplate.success(resDto, HttpStatus.OK));
 	}
 }
